Explicit casts for proto int32 to uint16_t

+++ src/lib/network/operator_translator.cpp
@@ -147,8 +147,8 @@ inline std::shared_ptr<OperatorTask> OperatorTranslator::translate(
   std::shared_ptr<JoinNestedLoopA> nested_loop_join;
 
   // TODO(anybody): joins without JOIN condition are currently not supported in this translator.
-  auto column_ids = std::make_pair(ColumnID{nested_loop_join_operator.left_column_id()},
-                                   ColumnID{nested_loop_join_operator.right_column_id()});
+  auto column_ids = std::make_pair(ColumnID{static_cast<uint16_t>(nested_loop_join_operator.left_column_id())},
+                                   ColumnID{static_cast<uint16_t>(nested_loop_join_operator.right_column_id())});
   auto join_columns = optional<std::pair<ColumnID, ColumnID>>(column_ids);
   nested_loop_join = std::make_shared<JoinNestedLoopA>(
       input_left_task->get_operator(), input_right_task->get_operator(), join_columns, scan_type, join_mode);
@@ -163,7 +163,7 @@ inline std::shared_ptr<OperatorTask> OperatorTranslator::translate(
 
 inline std::shared_ptr<OperatorTask> OperatorTranslator::translate(
     const proto::TableScanOperator& table_scan_operator) {
-  const auto column_id = ColumnID{table_scan_operator.column_id()};
+  const auto column_id = ColumnID{static_cast<uint16_t>(table_scan_operator.column_id())};
   const auto& scan_type = translate_scan_type(table_scan_operator.filter_operator());
   Assert((table_scan_operator.has_input_operator()), "Missing Input Operator in Table Scan.");
 
@@ -182,7 +182,7 @@ inline std::shared_ptr<OperatorTask> OperatorTranslator::translate(
 
 inline std::shared_ptr<OperatorTask> OperatorTranslator::translate(
     const proto::IndexColumnScanOperator& index_column_scan_operator) {
-  const auto column_id = ColumnID{index_column_scan_operator.column_id()};
+  const auto column_id = ColumnID{static_cast<uint16_t>(index_column_scan_operator.column_id())};
   const auto& scan_type = translate_scan_type(index_column_scan_operator.filter_operator());
   Assert((index_column_scan_operator.has_input_operator()), "Missing Input Operator in Index Column Scan.");
 
@@ -209,7 +209,7 @@ inline std::shared_ptr<OperatorTask> OperatorTranslator::translate(const proto::
 }
 
 inline std::shared_ptr<OperatorTask> OperatorTranslator::translate(const proto::SortOperator& sort_operator) {
-  const auto column_id = ColumnID{sort_operator.column_id()};
+  const auto column_id = ColumnID{static_cast<uint16_t>(sort_operator.column_id())};
   const auto ascending = sort_operator.ascending();
   const auto output_chunk_size = sort_operator.output_chunk_size();
   Assert((sort_operator.has_input_operator()), "Missing Input Operator in Sort.");
